requirements.txt
certifi==2021.5.30
chardet==4.0.0
docutils==0.17.1
idna==2.10
Kivy==2.0.0
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
numpy==1.21.0
pandas==1.2.5
Pillow==8.2.0
pkg-resources==0.0.0
py-slippi==1.6.1
py-ubjson==0.16.1
Pygments==2.9.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
termcolor==1.1.0
urllib3==1.26.6
